refactor: split workers execution_service into a per-implementation package (#2050)
